Refactor request classes
